Pad the segment arrays to expected length.
If the input data is missing samples the resulting arrays will be off.
Splitter will refuse to save the segments.

This happens in Tu-204 which internally uses 2s frames. Because we
support 64s or 4s boundary, the Tu-204 arrays may be off by 2s.

Instead of "fixing" the mechanics for one unusual frame we'd rather fix
the data by padding 2 "missing" seconds of data.